Add files via upload
This is a ZIP file detailing my portfolio project assignment for my Udacity Full Stack Developer Nanodegree program, featuring some of my other work in Python, as well as HTML. The site was built on Bootstrap 3 and project files are linked to my GitHub account (as you will see). Those files are also available to view.